op-challenger: Move generic split game logic to separate package.
